Fix overload resolution for useDeepState
It all started when `deepState` was a `CarpoolState`. `useDeepState`, `useDeepStateChangeHandler`, and `useDeepStateCheckChangeHandler` resolved to their generic overloads, which have return types of `any` instead of `T[K]`.

It turned out that this was because `CarpoolState` customizes `TChildrenStates`, which meant that it was not assignable to plain-old `DeepStateObject<T>`. It is, on the other hand, assignable to `DeepStateObject<T, any>`.

This change relaxes the type constraint on the `deepState` parameter by changing it to `DeepStateObject<T, any>`. Now, TypeScript selects the overload that returns `T[K]`, which is what we want.